% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrios.R
\docType{data}
\name{barrios}
\alias{barrios}
\title{Puerto Rico Barrios}
\format{
An object of class data.frame.
}
\usage{
barrios
}
\description{
Data set containing barrios per municipality. Note that the list is 
non-exhaustive and some barrios might be missing.
}
\details{
Data.frame contains 550 rows and 4 columns
\describe{
  \item{id}{Municipality name in lowercase.}
  \item{barrio}{Barrio name written as it might be found on a raw data set.}
  \item{original_barrio}{Barrio name correctly written.}
  \item{municipio}{Municipality name in capitalized format.}
}
}
\examples{

head(barrios)
}
\keyword{datasets}
