% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_helper_fncs.R
\name{extract_numeric_ranges}
\alias{extract_numeric_ranges}
\title{Helper Function 1: extract numeric values using regular expression}
\usage{
extract_numeric_ranges(text)
}
\arguments{
\item{text}{text from which to extract numeric values}
}
\value{
numeric values in the text
}
\description{
Helper Function 1: extract numeric values using regular expression
}
