% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_municipio.R
\name{fix_municipio}
\alias{fix_municipio}
\title{Fix municipio names}
\usage{
fix_municipio(
  x,
  min.dist = 2,
  min.score = 1/3,
  check.barrios = TRUE,
  min.score.barrio = 2/9,
  first.letter.match = TRUE,
  return.dist = FALSE,
  additional.barrios = NULL
)
}
\arguments{
\item{x}{character string or vector of character strings from Puerto Rico municipios}

\item{min.dist}{numeric.}

\item{min.score}{numeric. Refers to the minimum acceptable ... for municipio based matches.}

\item{check.barrios}{logical. Defaults to `TRUE`. If `FALSE` then the function does not verify if the input is a barrio.}

\item{min.score.barrio}{numeric. Refers to the minimum acceptable ... for barrio based matches. Note that this parameter is used when `check.barrios` is `TRUE`.}

\item{first.letter.match}{logical. Defaults to `TRUE`. If `FALSE` then the function does not use the first letter match.}

\item{return.dist}{logical. Defaults to `FALSE`. If `TRUE` returns the Levenshteins metric for each match.}

\item{additional.barrios}{Defaults to `NULL`. data.frame with the first column of additional barrios and the second column of their corresponding municipios.}
}
\description{
\code{fix_municipio} fixes typical data entry errors found in municipio names by matching each argument to the set correctly spelled municipios using Levenshtein metric.
}
\examples{

example <- c("franques", "manati", "Veja Baja" , "Comerip", "San Juan", "Mayaguez", "rio piedras")

fix_municipio(example)


}
