% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_estimates.R
\name{get_census_estimates}
\alias{get_census_estimates}
\title{Get tidy population estimates for Puerto Rico}
\usage{
get_census_estimates(years, product, municipio.pep = FALSE, census_key)
}
\arguments{
\item{years}{numeric argument. It refers to the years for which the user wants population estimates.}

\item{product}{character argument. Refers to the USCB product to call: "acs" for American Community Survey, "pep" for Population Estimates and Projecttions, and "decennial" for Decennial Census.}

\item{municipio.pep}{logical argument. Defaults to FALSE. If TRUE, it returns populations estimates (PEP) by municipio for preselected age groups. For ACS and Decennial products, municipio is always TRUE.}

\item{census_key}{character argument. The census key is provided by the US Census Bureau. To obtain a census key, please visit \url{https://api.census.gov/data/key_signup.html}.}
}
\description{
This function returns a tidy dataset of population estimates for Puerto Rico. It pulls data from the US Census Bureau population estimates and the  United Nations 2022 Revision of World Population Prospects (\url{https://population.un.org/wpp/}).
}
\examples{
\dontrun{
# Example usage of the function with sensitive information
get_census_estimates(2011:2015, product = "acs", municipio.pep = FALSE, census_key = yourcensuskey)
}
}
