% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tidy_pop_estimates.R
\name{make_tidy_pop_estimates}
\alias{make_tidy_pop_estimates}
\title{Make tidy pop estimates}
\usage{
make_tidy_pop_estimates(year, municipio = FALSE, un_data = FALSE, census_key)
}
\arguments{
\item{year}{numeric argument. It refers to the years for which the user wants population estimates. Earliest year is 1950 and can go beyond the actual year, in which case it provides UN population projections.}

\item{municipio}{logical argument. Defaults to FALSE. If TRUE, it returns populations estimates by municipio for preselected age groups.}

\item{un_data}{logical argument. Defaults to FALSE. If TRUE, it returns population estimates from the UN only. Otherwise it provides US Census Bureau population estimates when available.}

\item{census_key}{character argument. The census key is provided by the US Census Bureau. To obtain a census key, please visit [](https://api.census.gov/data/key_signup.html).}
}
\description{
This function returns a tidy dataset of population estimates for Puerto Rico. It pulls data from the US Census Bureau population estimates and the  United Nations [2022 Revision of World Population Prospects](https://population.un.org/wpp/).
}
