% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{map}
\alias{map}
\alias{map_centers}
\title{Puerto Rico Map}
\format{
An object of class data.frame.
}
\usage{
map
}
\description{
Data set containing data necessary to build a map for Puerto Rico, including 
its municipalities. It also includes `map_centers` which has municipalities 
and the approximate coordinate of the center.
}
\details{
map data frame contains 28,332 rows and 4 columns
\describe{
  \item{municipio}{Municipality name}
  \item{part}{...}
  \item{X}{\code{x} coordinate for polygon.}
  \item{Y}{\code{y} coordinate for polygon.}
}
}
\examples{

head(map)
head(map_centers)
}
\keyword{datasets}
