% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipio.abr.R
\docType{data}
\name{municipio.abr}
\alias{municipio.abr}
\title{Municipio Abbreviations Dataset}
\format{
A data frame with abbreviations for municipalities.
}
\source{
Information source (if applicable).
}
\usage{
municipio.abr
}
\description{
This dataset contains abbreviations for municipalities.
}
