% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipio.centers.R
\docType{data}
\name{municipio.centers}
\alias{municipio.centers}
\title{Municipio Centers Coordinates Dataset}
\format{
A data frame with center coordinates for municipalities.
}
\usage{
municipio.centers
}
\description{
This dataset coordinates for the center of each municipality
}
