% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipio.health.regions.R
\docType{data}
\name{municipio.health.regions}
\alias{municipio.health.regions}
\title{Municipio Health Regions Dataset}
\format{
A string vector with health region for each municipality.
}
\usage{
municipio.health.regions
}
\description{
This dataset contains the health region for each municipality. Sorted following the municipio.names object.
}
