% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipio.names.R
\docType{data}
\name{municipio.names}
\alias{municipio.names}
\title{Municipio Names Dataset}
\format{
A string vector with municipalities names.
}
\usage{
municipio.names
}
\description{
This dataset names for municipalities
}
