% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipios.R
\docType{data}
\name{municipios}
\alias{municipios}
\title{Municipio names, abbreviation and other useful objects}
\format{
municipio.abr, municipio.centers, municipio.health.regions, municipio.names
}
\usage{
municipios
}
\description{
\code{puertorico} currently contains the following "municipio" data sets. All data 
are arranged according to the alphabetical order of the municipality names.
}
\examples{
# Load the municipios dataset
municipios

}
\keyword{datasets}
