% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_helper_fncs.R
\name{rename_columns}
\alias{rename_columns}
\title{Helper Function 6: function to rename columns containing specific substrings}
\usage{
rename_columns(df, old_substrings, new_name)
}
\arguments{
\item{df}{data.frame/data.table}

\item{old_substrings}{old column names}

\item{new_name}{new column names}
}
\value{
data.table
}
\description{
Helper Function 6: function to rename columns containing specific substrings
}
