% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_helper_fncs.R
\name{vars_names}
\alias{vars_names}
\title{Helper Function 2: make a data.table of variable names}
\usage{
vars_names(list_var)
}
\arguments{
\item{list_var}{list from which to create a data.table}
}
\value{
data.table
}
\description{
Helper Function 2: make a data.table of variable names
}
