% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_helper_fncs.R
\name{wrangle_dat}
\alias{wrangle_dat}
\title{Helper Function 3: wrangle ACS and Decennial Census data}
\usage{
wrangle_dat(dat, product = NULL, variable_names = NULL, year_input)
}
\arguments{
\item{dat}{data as downloaded from USCB}

\item{product}{acs or decennial}

\item{variable_names}{variable data.table output from var_names function}

\item{year_input}{years in dat}
}
\value{
data.table
}
\description{
Helper Function 3: wrangle ACS and Decennial Census data
}
