% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_helper_fncs.R
\name{wrangle_pep}
\alias{wrangle_pep}
\title{Helper Function 4: wrangle PEP data}
\usage{
wrangle_pep(dat, municipio = NULL, year_input)
}
\arguments{
\item{dat}{data as downloaded from USCB}

\item{municipio}{TRUE or FALSE}

\item{year_input}{years in dat}
}
\value{
data.table
}
\description{
Helper Function 4: wrangle PEP data
}
