% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_english.R
\name{to_english}
\alias{to_english}
\title{To English}
\usage{
to_english(x)
}
\arguments{
\item{...}{Character argument.}
}
\description{
`to_english` takes an character argument and removes diacritic marks, dieresis or tildes from modern latin alphabet letters.
}
\details{
This function serves for English language character normalization by removing traditional symbols used in other languages.
}
